requirements.txt
flask==3.1.1
flask-cors==6.0.2
websockets==15.0.1
python-dotenv==1.1.0
pymongo==4.15.5
dnspython==2.8.0
google-generativeai==0.8.6
easyocr==1.7.2
deep-translator==1.11.4
pillow==10.4.0
numpy==2.2.6
gunicorn==21.2.0
